Create Exercise9_Divider.java
This program is almost the same as Exercise 8's. However, this exercise teaches the coder how an int variable value should be converted into a double value for correct division. For example, computers see 5/2 = 2. We know from math schooling and our calculators will correctly tell us 5/2 = 2.5, which is the result we get from this program as "num1" is converted from, let's say 6...to 6.0. 6.0 is a proper double variable value.